Arduino Code*
Arduino Code is changed acording to the previously changed Raspberry pi code. The pin readings are done, WITHOUT the PID control. PID control will be done according to the error read. What code does is the following:

error is the variable reading the 3 bits and forming the absolute error. In the loop, if distance OK, the next command is to get whether error is on the right or left, then a code "pwmcontrol " will be called, which will get the global variable error, make a PID control and output two values for the speeds of the motors. There will be two integers for two duty cysles, one low other large. And these two can be inter changed between two motors according to the error's position. This is for the simplicity of the code.